requirements.txt
# SoundPixel-AI — Python dependencies
# Run: pip install -r requirements.txt

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.7
pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
opencv-python>=4.8.0